#21 test: user_spec.rb集約
